requirements.txt
alt-profanity-check==1.6.1
better-profanity==0.7.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.4
peewee==3.17.9
scikit-learn==1.6.1
scipy==1.15.2
threadpoolctl==3.6.0
Werkzeug==3.1.3
